put continues back

+++ plugin_parser.py
@@ -156,12 +156,14 @@ def add_plugins(path_dirs, plugin_url):
         for subdir in subdirs:
             if os.path.isdir(subdir):
                 directory = subdir.split(path_dirs.plugin_repos+"/"+plugin_name+"/")[0]
+                if subdir == path_dirs.plugin_repos+"/"+plugin_name:
+                    continue
                 if not os.path.isdir(path_dirs.base_dir + directory):
                     print "Failed to install "+plugin_name+" resource: "+directory
                     os.system("sudo rm -rf "+path_dirs.plugin_repos+"/"+plugin_name)
                     return
         # resources installed correctly. Building...
-        os.system("/bin/sh /data/build_images.sh")
+        os.system("/bin/sh "+path_dirs.base_dir+"build_images.sh")
     except Exception as e:
         pass
 
@@ -227,6 +229,7 @@ def remove_plugins(path_dirs, plugin_url):
                                 with open(path_dirs.template_dir + namespace + ".template", 'w') as configfile:
                                     config.write(configfile)
                                 shutil.rmtree(s_sub)
+                continue
             elif plugin_name+"/visualization" in r_sub:
                 names = os.listdir(path_dirs.vis_dir)
                 has_subdir = False
@@ -240,6 +243,7 @@ def remove_plugins(path_dirs, plugin_url):
                     for dirpath, dirnames, files in os.walk(path_dirs.vis_dir+"/"):
                         for file in files:
                             os.remove(path_dirs.vis_dir+"/"+file)
+                    continue
                 sys_subdirs = [x[0] for x in os.walk(path_dirs.vis_dir + "/")]
                 repo_dir = r_sub.split(plugin_name+"/visualization/")[-1]
                 namespace = "visualization"
@@ -255,9 +259,12 @@ def remove_plugins(path_dirs, plugin_url):
                     for dirpath, dirnames, files in os.walk(path_dirs.collectors_dir + "/"):
                         for file in files:
                             os.remove(path_dirs.collectors_dir+"/"+file)
+                    continue
                 sys_subdirs = [x[0] for x in os.walk(path_dirs.collectors_dir + "/")]
                 repo_dir = r_sub.split(plugin_name+"/collectors/")[-1]
                 namespace = "collectors"
+            else:
+                continue
             #handles case if item removed is not in plugins/
             for s_sub in sys_subdirs:
                 if repo_dir in s_sub:
